🐛 5ZKP6T workflow: split foreign replica routing
