+++ .agentplane/tasks/202607252235-5ZKP6T/pr/meta.json
@@ -2,9 +2,9 @@
   "base": "main",
   "branch": "task/202607252235-5ZKP6T/prevent-foreign-task-artifacts-in-branch-pr-work",
   "created_at": "2026-07-25T22:39:31.309Z",
-  "diffstat_sha256": "sha256:1e4bef84d13a2d64aa05ddeee595346ed89d67e0e79572ed73062d45827d1210",
-  "last_verified_at": "2026-07-26T01:13:25.790Z",
-  "last_verified_diffstat_sha256": "sha256:4d862db043be0bf0129242273c681c34cf52bfecd931e912715605f697e481f5",
+  "diffstat_sha256": "sha256:e398ff4c922848020a68f53d82c4df3570d786ae9309a259638d8a41aa3f9bc7",
+  "last_verified_at": "2026-07-26T01:53:14.289Z",
+  "last_verified_diffstat_sha256": "sha256:1e4bef84d13a2d64aa05ddeee595346ed89d67e0e79572ed73062d45827d1210",
   "pr_number": 4625,
   "pr_url": "https://github.com/basilisk-labs/agentplane/pull/4625",
   "schema_version": 1,
@@ -12,6 +12,6 @@
   "task_id": "202607252235-5ZKP6T",
   "updated_at": "2026-07-25T22:40:26.297Z",
   "verify": {
-    "status": "fail"
+    "status": "pass"
   }
 }
